+++ src/services/contentAggregator.js
@@ -1,7 +1,12 @@
-import { getProgressStateByIds, getProgressPercentageByIds, getResumeTimeSecondsByIds } from "./contentProgress"
-import { isContentLikedByIds } from "./contentLikes"
-import { fetchLikeCount, fetchLastInteractedChild } from "./railcontent"
-
+import {
+  getLastInteractedOf,
+  getNextLesson,
+  getProgressPercentageByIds,
+  getProgressStateByIds,
+  getResumeTimeSecondsByIds
+} from "./contentProgress"
+import {isContentLikedByIds} from "./contentLikes"
+import {fetchLastInteractedChild, fetchLikeCount} from "./railcontent"
 
 
 export async function addContextToContent(dataPromise, ...dataArgs)
@@ -18,6 +23,8 @@ export async function addContextToContent(dataPromise, ...dataArgs)
     addProgressStatus = false,
     addResumeTimeSeconds = false,
     addLastInteractedChild = false,
+    addNextLesson = false,
+    addLastInteractedParent = false,
   } = options
 
   const dataParam = lastArg === options ? dataArgs.slice(0, -1) : dataArgs
@@ -25,7 +32,8 @@ export async function addContextToContent(dataPromise, ...dataArgs)
   const data = await dataPromise(...dataParam)
   if(!data) return false
 
-  let ids = []
+  let items = []
+  let dataMap = []
 
   if (dataField && (data?.[dataField] || iterateDataFieldOnEachArrayElement)) {
     if (iterateDataFieldOnEachArrayElement && Array.isArray(data)) {
@@ -36,19 +44,35 @@ export async function addContextToContent(dataPromise, ...dataArgs)
       ids = data[dataField].map(item => item?.id).filter(Boolean);
     }
   } else if (Array.isArray(data)) {
-    ids = data.map(item => item?.id).filter(Boolean);
+    items = data;
   } else if (data?.id) {
-    ids = [data.id]
+    items = [data]
+  }
+
+  const ids = items.map(item => item?.id).filter(Boolean)
+
+  //create data structure for common use by functions
+  if (addNextLesson) {
+    items.forEach((item) => {
+      if (item?.id) {
+        dataMap.push({
+          'children': item.children?.map(child => child.id) ?? [],
+          'type': item.type,
+          'id': item.id,
+        })
+      }
+    })
   }
 
   if(ids.length === 0) return false
 
-  const [progressPercentageData, progressStatusData, isLikedData, resumeTimeData, lastInteractedChildData] = await Promise.all([
+  const [progressPercentageData, progressStatusData, isLikedData, resumeTimeData, lastInteractedChildData, nextLessonData] = await Promise.all([
     addProgressPercentage ? getProgressPercentageByIds(ids) : Promise.resolve(null),
     addProgressStatus ? getProgressStateByIds(ids) : Promise.resolve(null),
     addIsLiked ? isContentLikedByIds(ids) : Promise.resolve(null),
     addResumeTimeSeconds ? getResumeTimeSecondsByIds(ids) : Promise.resolve(null),
     addLastInteractedChild ? fetchLastInteractedChild(ids)  : Promise.resolve(null),
+    (addNextLesson || addLastInteractedParent) ? getNextLesson(dataMap) : Promise.resolve(null),
   ])
 
   const addContext = async (item) => ({
@@ -59,7 +83,13 @@ export async function addContextToContent(dataPromise, ...dataArgs)
     ...(addLikeCount && ids.length === 1 ? { likeCount: await fetchLikeCount(item.id) } : {}),
     ...(addResumeTimeSeconds ? { resumeTime: resumeTimeData?.[item.id] } : {}),
     ...(addLastInteractedChild ? { lastInteractedChild: lastInteractedChildData?.[item.id] } : {}),
+    ...(addNextLesson ? { nextLesson: nextLessonData?.[item.id] } : {}),
   })
+
+  if (addLastInteractedParent) {
+    const parentId = await getLastInteractedOf(ids);
+    data['nextLesson'] = nextLessonData[parentId];
+  }
 
   if (dataField) {
     if (iterateDataFieldOnEachArrayElement) {